Make fts engineering idempotant
Only calculate features if it has not yet been done. This allows to
just rerun the full slurm pipeline job, if a subset of cities failed.

+++ eubucco/ft_eng/ft_eng.py
@@ -51,27 +51,27 @@ def create_features(city_path,
 
     city_fts = pd.DataFrame(index=[0])
 
-    if bld:
+    if bld and not os.path.isfile(paths['bld_fts']):
         building_fts = buildings.merge(ufo_buildings.features_building_level(buildings), on='id')
         building_fts.loc[indexes_].drop(columns=['geometry']).to_csv(paths['bld_fts'], index=False)
         city_fts = pd.concat((city_fts, ufo_city.features_city_level_buildings(buildings.loc[indexes_])), axis=1)
 
-    if bld_d:
+    if bld_d and not os.path.isfile(paths['bld_d_fts']):
         building_fts = ufo_buildings.features_buildings_distance_based(buildings, building_fts)
         building_fts.loc[indexes_].to_csv(paths['bld_d_fts'], index=False)
 
-    if blk:
+    if blk and not os.path.isfile(paths['block_fts']):
         block_fts = buildings.merge(ufo_blocks.features_block_level(buildings), on='id')
         block_fts.loc[indexes_].drop(columns=['geometry']).to_csv(paths['block_fts'], index=False)
         city_fts = pd.concat((city_fts, ufo_city.features_city_level_blocks(block_fts.loc[indexes_])), axis=1)
 
-    if blk_d:
+    if blk_d and not os.path.isfile(paths['block_d_fts']):
         block_fts = ufo_blocks.features_blocks_distance_based(buildings, block_fts)
         block_fts.loc[indexes_].to_csv(paths['block_d_fts'], index=False)
 
     buildings = buildings.loc[indexes_]
 
-    if int_:
+    if int_ and not os.path.isfile(paths['int_fts']):
         intersections = ufo_helpers.import_csv_w_wkt_to_gdf(paths['intersections'], CRS_UNI)
         fts_int = buildings[['id']]
         fts_int['dist_to_closest_int'] = ufo_streets.feature_distance_to_closest_intersection(
@@ -81,7 +81,7 @@ def create_features(city_path,
         fts_int.to_csv(paths['int_fts'], index=False)
         city_fts = pd.concat((city_fts, ufo_city.feature_city_level_intersections(intersections)), axis=1)
 
-    if str_:
+    if str_ and not os.path.isfile(paths['str_fts']):
         streets = ufo_helpers.import_csv_w_wkt_to_gdf(paths['streets'], CRS_UNI)
         fts_str = buildings[['id']]
         fts_str = pd.concat((fts_str, ufo_streets.features_closest_street(buildings, streets)), axis=1)
@@ -89,7 +89,7 @@ def create_features(city_path,
         fts_str.to_csv(paths['str_fts'], index=False)
         city_fts = pd.concat((city_fts, ufo_city.features_city_level_streets(streets)), axis=1)
 
-    if sbb_:
+    if sbb_ and not os.path.isfile(paths['sbb_fts']):
         sbb = ufo_helpers.import_csv_w_wkt_to_gdf(paths['sbb'], CRS_UNI)
         fts_sbb = buildings[['id']]
         fts_sbb = pd.concat((fts_sbb, ufo_streets.features_own_sbb(buildings, sbb)), axis=1)
@@ -97,7 +97,7 @@ def create_features(city_path,
         fts_sbb.to_csv(paths['sbb_fts'], index=False)
         city_fts = pd.concat((city_fts, ufo_city.features_city_level_sbb(sbb)), axis=1)
 
-    if city_level:
+    if city_level and not os.path.isfile(paths['city_level_fts']):
         print('Generating city-level features...')
         city_fts.to_csv(paths['city_level_fts'], index=False)
